fix windows compile warnings

+++ vowpalwabbit/boosting.cc
@@ -254,7 +254,7 @@ void save_load_sampling(boosting &o, io_buf &model_file, bool read, bool text)
   stringstream os;
   os << "boosts " << o.N << endl;
   const char* buff = os.str().c_str();
-  bin_text_read_write_fixed(model_file, (char *) &(o.N),  sizeof(o.N), "", read, buff, strlen(buff), text);
+  bin_text_read_write_fixed(model_file, (char *) &(o.N),  sizeof(o.N), "", read, buff, (uint32_t)strlen(buff), text);
 
   if (read)
   { o.alpha.resize(o.N);
@@ -271,7 +271,7 @@ void save_load_sampling(boosting &o, io_buf &model_file, bool read, bool text)
     { stringstream os2;
       os2 << "alpha " << o.alpha[i] << endl;
       const char* buff2 = os.str().c_str();
-      bin_text_write_fixed(model_file, (char *) &(o.alpha[i]),  sizeof(o.alpha[i]), buff2, strlen(buff2), text);
+      bin_text_write_fixed(model_file, (char *) &(o.alpha[i]),  sizeof(o.alpha[i]), buff2, (uint32_t)strlen(buff2), text);
     }
 
   for (int i = 0; i < o.N; i++)
@@ -284,7 +284,7 @@ void save_load_sampling(boosting &o, io_buf &model_file, bool read, bool text)
     { stringstream os2;
       os2 << "v " << o.v[i] << endl;
       const char* buff2 = os.str().c_str();
-      bin_text_write_fixed(model_file, (char *) &(o.v[i]),  sizeof(o.v[i]), buff2, strlen(buff2), text);
+      bin_text_write_fixed(model_file, (char *) &(o.v[i]),  sizeof(o.v[i]), buff2, (uint32_t)strlen(buff2), text);
     }
 
   if (read)
@@ -316,7 +316,7 @@ void save_load(boosting &o, io_buf &model_file, bool read, bool text)
   stringstream os;
   os << "boosts " << o.N << endl;
   const char* buff = os.str().c_str();
-  bin_text_read_write_fixed(model_file, (char *) &(o.N),  sizeof(o.N), "", read, buff, strlen(buff), text);
+  bin_text_read_write_fixed(model_file, (char *) &(o.N),  sizeof(o.N), "", read, buff, (uint32_t)strlen(buff), text);
 
   if (read)
   { o.alpha.resize(o.N);
@@ -332,7 +332,7 @@ void save_load(boosting &o, io_buf &model_file, bool read, bool text)
     { stringstream os2;
       os2 << "alpha " << o.alpha[i] << endl;
       const char* buff2 = os.str().c_str();
-      bin_text_write_fixed(model_file, (char *) &(o.alpha[i]),  sizeof(o.alpha[i]), buff2, strlen(buff2), text);
+      bin_text_write_fixed(model_file, (char *) &(o.alpha[i]),  sizeof(o.alpha[i]), buff2, (uint32_t)strlen(buff2), text);
     }
 
   if (read)
